Working mavlink router when hub is run without docker
Only message forwarding works in docker-compose (messages can't be sent both ways with hub). Does work when running hub outside of the compose with the --network=host flag and setting the MAV_OUTPUT IP to localhost/127.0.0.1. InfluxDB check is commented out temporarily for testing without the rest of the docker-compose containers