working mavlink routing/parsing
previously the router was spawned with a new goroutine inside RunMavlink
but the RunMavlink goroutine would finish right after the router was
started. RunMavlink used to have a for loop at the end to check for new
path planning paths. Now we commented it out and RunMavlink won't keep
running.