fix: QGC can now arm/disarm plane and upload flight plans to SITL
Previously, when using QGroundControl with the mavlink router of hub we
were not able to arm/disarm the plane and upload flight data to the
SITL. QGC was trying to send mavlink packets over UDP to hub but for
some reason was unable to with the default docker-compose settings.
Now the docker containers run in host mode to resolve this issue. More
info here: https://app.clickup.com/t/2cdb9yd